backend/requirements.txt
aiosmtplib==5.1.0
PyPDF2==3.0.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
fastapi-mail==1.6.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kaggle==2.0.0
kagglesdk==0.1.16
MarkupSafe==3.0.3
motor==3.7.1
packaging==26.0
protobuf==7.34.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-http-client==3.3.7
python-jose==3.5.0
python-slugify==8.0.4
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
sendgrid==6.12.5
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
Werkzeug==3.1.6
